+++ src/components/FlawFormOwner.vue
@@ -9,7 +9,7 @@ function selfAssign() {
   owner.value = userStore.userName;
 }
 
-function handleClick(fn: (arg?: any) => any){
+function handleClick(fn: (arg?: any) => any) {
   selfAssign();
   nextTick(fn);
 }
@@ -22,26 +22,18 @@ const isAssignedToMe = computed(() => owner.value === userStore.userName);
   <label class="ps-3 mb-3 input-group osim-input">
     <div class="row">
       <span class="form-label col-3 pe-3 ">
-        Assignee
+        Owner
       </span>
-      <EditableText
-        v-model="owner"
-        label="Assignee"
-        type="text"
-      >
+      <EditableText v-model="owner">
         <template v-if="!isAssignedToMe" #buttons-out-of-editing-mode="{ onBlur }">
-          <button
-            type="button"
-            class="btn btn-primary osim-self-assign"
-            @click.prevent.stop="handleClick(onBlur)"
-          >Self Assign</button>
+          <button type="button" class="btn btn-primary osim-self-assign" @click.prevent.stop="handleClick(onBlur)">
+            Self Assign
+          </button>
         </template>
         <template v-if="!isAssignedToMe" #buttons-in-editing-mode="{ onBlur }">
-          <button
-            type="button"
-            class="btn btn-primary osim-self-assign"
-            @click.prevent.stop="handleClick(onBlur)"
-          >Self Assign</button>
+          <button type="button" class="btn btn-primary osim-self-assign" @click.prevent.stop="handleClick(onBlur)">
+            Self Assign
+          </button>
         </template>
       </EditableText>
     </div>